requirements.txt
django==4.2.9
djangorestframework==3.14.0
beautifulsoup4==4.12.2
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
groq==0.4.0
pandas==2.0.3
pytest==7.4.3
black==23.11.0
isort==5.12.0
python-decouple==3.8
